+++ src/main/scala/com/lucendar/common/serv/utils/ServUtils.scala
@@ -7,8 +7,12 @@
  * ***************************************************************************** */
 package com.lucendar.common.serv.utils
 
+import com.google.common.base.Throwables
+import info.gratour.common.error.Errors
 import org.apache.commons.codec.binary.Base64
+import org.apache.commons.io.FilenameUtils
 import org.springframework.core.io.ResourceLoader
+import org.springframework.http.{HttpStatus, MediaType}
 
 import java.io.{IOException, InputStream}
 import java.nio.ByteBuffer
@@ -112,4 +116,129 @@ object ServUtils {
     r.stream
   }
 
+  final val MimeTypeDefault = "application/octet-stream"
+  final val MediaTypeDefault = MediaType.parseMediaType(MimeTypeDefault)
+
+  final val FileExtMimeMap: Map[String, String] = Seq(
+    "txt" -> "text/plain",
+
+    // video
+    "mp4" -> "video/mp4",
+    "mpeg" -> "video/mpeg",
+    "ogv" -> "video/ogg",
+    "avi" -> "video/x-msvideo",
+    "ts" -> "video/mp2t",
+    "webm" -> "video/webm",
+
+    // audio
+    "wav" -> "audio/wav",
+    "oga" -> "audio/ogg",
+    "mp3" -> "audio/mpeg",
+    "aac" -> "audio/aac",
+    "opus" -> "audio/opus",
+    "weba" -> "audio/webm",
+
+
+
+    // image
+    "png" -> "image/png",
+    "jpg" -> "image/jpg",
+    "jpeg" -> "image/jpg",
+    "gif" -> "image/gif",
+    "bmp" -> "image/bmp",
+    "webp" -> "image/webp",
+    "tif" -> "image/tiff",
+    "tiff" -> "image/tiff",
+
+    // html
+    "htm" -> "text/html",
+    "html" -> "text/html",
+    "css" -> "text/css",
+    "js" -> "text/javascript",
+    "mjs" -> "text/javascript",
+    "ttf" -> "font/ttf",
+    "woff" -> "font/woff",
+    "woff2" -> "font/woff2",
+    "otf" -> "font/otf",
+
+    // document
+    "pdf" -> "application/pdf",
+    "doc" -> "application/msword",
+    "docx" -> "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
+    "xls" -> "application/vnd.ms-excel",
+    "xlsx" -> "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
+    "ppt" -> "application/vnd.ms-powerpoint",
+    "pptx" -> "application/vnd.openxmlformats-officedocument.presentationml.presentation",
+
+    // data
+    "json" -> "application/json",
+    "xml" -> "application/xml",
+
+    // archive
+    "zip" -> "application/zip",
+    "7z" -> "application/x-7z-compressed",
+    "jar" -> "application/java-archive",
+
+    "bin" -> "application/octet-stream"
+  ).toMap
+
+  val FileExtMediaTypeMap: Map[String, MediaType] = FileExtMimeMap
+    .map(a => {
+      (a._1, MediaType.parseMediaType(a._2))
+    })
+
+  def mapMediaTypeFromFileNameDefault(fileName: String): MediaType = {
+    val ext = FilenameUtils.getExtension(fileName).toLowerCase
+    FileExtMediaTypeMap.getOrElse(ext, MediaTypeDefault)
+  }
+
+  private def messageOf(t: Throwable): String =
+    if (t != null) {
+      val m = t.getMessage
+      if (m != null)
+        m
+      else
+        ""
+    } else
+      ""
+
+  def rootMessageOf(t: Throwable): String =
+    if (t != null)
+      messageOf(Throwables.getRootCause(t))
+    else
+      ""
+
+  def errCodeToHttpStatusCode(errCode: Int): Int = {
+    errCode match {
+      case Errors.OK =>
+        HttpStatus.OK.value()
+
+      case Errors.ACCESS_DENIED | Errors.NOT_ENOUGH_PRIV =>
+        HttpStatus.FORBIDDEN.value()
+
+      case Errors.NOT_AUTHENTICATED | Errors.AUTHENTICATION_FAILED | Errors.INVALID_TOKEN | Errors.SESSION_EXPIRED =>
+        HttpStatus.UNAUTHORIZED.value()
+
+      case Errors.INTERNAL_ERROR =>
+        HttpStatus.INTERNAL_SERVER_ERROR.value()
+
+      case Errors.SERVICE_UNAVAILABLE | Errors.SERVICE_BUSY =>
+        HttpStatus.SERVICE_UNAVAILABLE.value()
+
+      case Errors.RECORD_NOT_FOUND | Errors.FILE_NOT_FOUND =>
+        HttpStatus.NOT_FOUND.value()
+
+      case Errors.UNSUPPORTED_TYPE =>
+        HttpStatus.NOT_IMPLEMENTED.value()
+
+      case Errors.RESOURCE_OCCUPIED | Errors.TOO_MANY_REQUEST =>
+        HttpStatus.TOO_MANY_REQUESTS.value()
+
+      case Errors.TIMEOUT =>
+        HttpStatus.REQUEST_TIMEOUT.value()
+
+      case _ =>
+        HttpStatus.BAD_REQUEST.value()
+    }
+  }
 }